Implement acceleration structures and textures; add HW8/HW9 logic

+++ part2/camera.cpp
@@ -41,7 +41,7 @@ void Camera::Set_Resolution(const ivec2& number_pixels_input)
 }
 
 // Find the world position of the input pixel
-vec3 Camera::World_Position(const ivec2& pixel_index)
+vec3 Camera::World_Position(const vec2& pixel_index) // Changed ivec2 to vec2
 {
     vec2 C = Cell_Center(pixel_index);
     vec3 result = film_position + (C[0] * horizontal_vector) + (C[1] * vertical_vector);
