+++ part2/camera.h
@@ -35,7 +35,7 @@ public:
     void Set_Resolution(const ivec2& number_pixels_input);
 
     // Used for determining the where pixels are
-    vec3 World_Position(const ivec2& pixel_index);
+vec3 World_Position(const vec2& pixel_index);
     vec2 Cell_Center(const ivec2& index) const
     {
         return min+(vec2(index)+vec2(.5,.5))*pixel_size;
